Avoid nested components with a top-level `renderResult`
