+++ frontend/src/employee-mobile-frontend/App.tsx
@@ -51,7 +51,7 @@ import { UnitOrGroup, toUnitOrGroup } from './common/unit-or-group'
 import MessagesPage from './messages/MessagesPage'
 import NewChildMessagePage from './messages/NewChildMessagePage'
 import NewMessagePage from './messages/NewMessagePage'
-import ReceivedThreadPage from './messages/ReceivedThreadPage'
+import { ReceivedThreadPage } from './messages/ThreadView'
 import { UnreadMessagesPage } from './messages/UnreadMessagesPage'
 import { MessageContextProvider } from './messages/state'
 import MobileLander from './pairing/MobileLander'
